size?

+++ src/eicQuickSim/Weights.C
@@ -294,11 +294,11 @@ void Weights::loadPrecalculatedWeights(const std::string &precalcCSVFilename) {
 double Weights::getWeight(double Q2) const {
     int idx = -1;
 
-    for (size_t i = 0; i < Q2mins.size(); i++) {
+    for (int i = 0; i < Q2mins.size(); i++) {
         std::cout << "A" << std::endl;
         if (inQ2Range(Q2, Q2mins[i], Q2maxs[i], false)) {
             std::cout << "HERE" << std::endl;
-            idx = static_cast<int>(i);
+            idx = i;
         }
         std::cout << idx << std::endl;
     }
